chore: let tbs depend on core, use derive to implement encodable / decodable on its public types
