+++ fedimint-core/Cargo.toml
@@ -33,7 +33,6 @@ strum = { workspace = true }
 strum_macros = { workspace = true }
 hex = { version = "0.4.3", features = [ "serde"] }
 sha3 = "0.10.8"
-tbs = { package = "fedimint-tbs", version = "0.3.0-alpha", path = "../crypto/tbs" }
 tokio = { version = "1.36.0", features = ["sync", "io-util"] }
 thiserror = { workspace = true }
 tracing = { workspace = true }
